refactor: 로그인 인터페이스로 구성
