+++ src/main/java/com/seoridam/rehearserver/controller/MemberController.java
@@ -4,6 +4,7 @@ import com.seoridam.rehearserver.dto.JoinRequestDto;
 import com.seoridam.rehearserver.dto.LoginRequestDto;
 import com.seoridam.rehearserver.global.common.StatusEnum;
 import com.seoridam.rehearserver.global.common.SuccessResponse;
+import com.seoridam.rehearserver.service.LoginService;
 import com.seoridam.rehearserver.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -18,6 +19,7 @@ import static com.seoridam.rehearserver.global.common.StatusEnum.OK;
 public class MemberController {
 
     private final MemberService memberService;
+    private final LoginService loginService;
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("/join")
@@ -40,7 +42,7 @@ public class MemberController {
 
     @PostMapping("/login")
     public SuccessResponse login(@Valid @RequestBody final LoginRequestDto loginRequestDto) {
-        String token = memberService.login(loginRequestDto);
+        String token = loginService.login(loginRequestDto);
         return SuccessResponse.builder()
                 .status(OK)
                 .message("로그인 성공")
