+++ src/main/java/com/seoridam/rehearserver/service/JwtLoginService.java
@@ -1,0 +1,28 @@
+package com.seoridam.rehearserver.service;
+
+import com.seoridam.rehearserver.domain.Member;
+import com.seoridam.rehearserver.dto.LoginRequestDto;
+import com.seoridam.rehearserver.global.util.JwtTokenProvider;
+import com.seoridam.rehearserver.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+
+
+@Service
+@RequiredArgsConstructor
+public class JwtLoginService implements LoginService{
+
+    private final MemberRepository memberRepository;
+    private final PasswordEncoder passwordEncoder;
+    private final JwtTokenProvider jwtTokenProvider;
+
+    public String login(LoginRequestDto loginRequestDto){
+        Member member = memberRepository.findByEmail(loginRequestDto.getEmail())
+                .orElseThrow(() -> new IllegalArgumentException("가입되지 않은 계정입니다."));
+        if (!passwordEncoder.matches(loginRequestDto.getPassword(), member.getPassword())) {
+            throw new IllegalArgumentException("잘못된 비밀번호입니다.");
+        }
+        return jwtTokenProvider.createToken(member.getUsername(), member.getRoles());
+    }
+}
