+++ src/main/java/com/seoridam/rehearserver/service/LoginService.java
@@ -1,0 +1,8 @@
+package com.seoridam.rehearserver.service;
+
+import com.seoridam.rehearserver.dto.LoginRequestDto;
+
+public interface LoginService {
+
+    String login(LoginRequestDto loginRequestDto);
+}
