+++ src/main/java/com/seoridam/rehearserver/service/MemberService.java
@@ -2,8 +2,6 @@ package com.seoridam.rehearserver.service;
 
 import com.seoridam.rehearserver.domain.Member;
 import com.seoridam.rehearserver.dto.JoinRequestDto;
-import com.seoridam.rehearserver.dto.LoginRequestDto;
-import com.seoridam.rehearserver.global.util.JwtTokenProvider;
 import com.seoridam.rehearserver.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -11,13 +9,12 @@ import org.springframework.stereotype.Service;
 
 import java.util.Collections;
 
-@Service
 @RequiredArgsConstructor
+@Service
 public class MemberService {
 
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
-    private final JwtTokenProvider jwtTokenProvider;
 
     public void join(JoinRequestDto dto){
         if(!dto.getPassword().equals(dto.getCheckPassword()))
@@ -56,14 +53,5 @@ public class MemberService {
 
     public boolean isDuplicatedNickname(String nickname){
         return memberRepository.findByNickname(nickname).isPresent();
-    }
-
-    public String login(LoginRequestDto loginRequestDto){
-        Member member = memberRepository.findByEmail(loginRequestDto.getEmail())
-                .orElseThrow(() -> new IllegalArgumentException("가입되지 않은 계정입니다."));
-        if (!passwordEncoder.matches(loginRequestDto.getPassword(), member.getPassword())) {
-            throw new IllegalArgumentException("잘못된 비밀번호입니다.");
-        }
-        return jwtTokenProvider.createToken(member.getUsername(), member.getRoles());
     }
 }
